SPARK-27964: Move v2 catalog update methods to CatalogV2Util.
